requirements.txt
mcp>=1.0.0
httpx>=0.27,<1.0
pydantic>=2.9,<3.0
structlog>=24.0,<25.0
paramiko>=3.0,<4.0
PyYAML>=6.0,<7.0
pytest>=8.3,<9.0
pytest-asyncio>=0.24,<1.0
